extracted block processing function
